Add edge check
Add a check to see if the trailed source beginning or end pixels are off chip. If so, flag them. Otherwise, check if the beginning or end pixel of a trail lands on a pixel marked with an edge flag. If so, add and edge flag to those trailed sources.